Remove space causing error.
When I ran the sample of create-table it threw an error. After researching there was a space between ReadCapacityUnites=10, and WriteCapacityUnites=5. I removed the space and the create-table is working now. This Pull Request matches the documentation for the same example (that works) from this page: https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Tools.CLI.html