Update GSI.md
Removed bad example of additional projected attributes. This deleted example is both confusing and confused.

Confusing: it refers back to a version of GameTitleIndex that does not have the UserId attribute projected into it. No such version exists.

Confused: Projecting the Wins and Losses attributes into GameTitleIndex does not make querying for highest ratio of wins to losses efficient because Wins are Losses are not part of the sort key. To get that ratio, you have to scan the entire index (or at least all entries for a particular game) and do the calculations manually. This could be done just as efficiently on the base table itself.

The only benefit of projecting Wins and Losses is the ability is to *display* those attributes along side the top scorer without needing to go back and do an additional Query on the base table. Or, the ability to display the list of top scorers for a particular game sorted by win/loss percentage (doing the calculation and sorting manually on the small number of top scorers) without having to do the additional Query on the base table.

I guess if the number of attributes in the base table was very large, doing the full scan on an index with only a few attributes is an efficiency in term of provisioned reads, but that's not the gist of the explanation in the text.