clarify that the limit applies to the items read
The previous wording did not seem to do enough to explain the difference in how DynamoDB's limit works when compared to a traditional SQL limit. In DynamoDB, you are limiting the number of items scanned / read, not the number of items returned (of course, you'll never return more items than were read, so it _is_ limiting the number of items returned, but only as a side effect).