Correct docs for `contains` function operand
Previously, the docs for the `contains` function said:

> In either case, `operand` must be a `String`.

This is incorrect. If the attribute specified by the `path` is a set, the `operand` must be the same type as the set's elements.  I had a binary set and thought this meant I had to pass my binary as a string to check if the binary set contains my binary value, but based on testing I have done, it works just fine to pass the operand as a binary in my case.