Added a note on native backup and restore feature
The native Backup and restore feature of DynamoDB is easy and efficient than using AWS Data Pipeline. As of now (Aug, 2020) the only reason, we should use Data Pipeline is when we want to restore table in a different AWS account. 
Data pipeline consumes read capacity from the base table which might impact a production table and you need to do additional capacity planning for EMR data nodes for this solution.