Update GettingStarted.NET.01.md
Bug fixed.
P. S. All the sample code in this tutorial is hard to read. Might be worth getting someone to clean this up and make it readable and use practices that are widely used in C#, and not a big deviation from it: variable names, method names, use of brackets, avoiding state in classes, avoiding static variables, using variables that are explained in the tutorial (without downloading your sample code, the tutorial makes little sense), etc.