Update V2globaltables_reqs_bestpractices.md
The current language is ambiguous. "In this case, you should expect to consume 10 rWCUs (or 10 replicated write request units, if using on-demand capacity) in both Ohio and N. Virginia" could mean two things: 1/ You consume 10 rWCU in each of Ohio and N. Virginia and 20 rWCU in total. 2/ It could mean that you consume 10 rWCU in both Ohio and N. Virginia combined. The change removes the ambiguity.