Update howitworks-ttl.md
I think you should consider adding an "and" to line #5. Doing so makes the sentence a little more clear. Without the "and" the phrase "the attribute’s value is a timestamp" could be interpreted as fact and not a condition that has to be met. 
Thank you for taking the time to review my proposed change.