Change "table" to "bucket", DataExport.HowItWorks.md
It says that "You can export data to an S3 bucket owned by a different account if you have the correct permissions to write to that table". Did it mean to say "bucket" (the S3 bucket in a different account) instead of "table" (supposedly a DynamoDB table)?

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.